+++ src/vs/v1_20_5/main/java/com/owlmaddie/datagen/CreatureChatLangProvider.java
@@ -7,14 +7,14 @@ import com.owlmaddie.chat.Advancements;
 import com.owlmaddie.chat.EntityChatData;
 import com.owlmaddie.i18n.CCText;
 import com.owlmaddie.utils.Randomizer;
+
+import java.util.Map;
+import java.util.TreeMap;
+import java.util.concurrent.CompletableFuture;
+import java.util.stream.Stream;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricLanguageProvider;
 import net.minecraft.core.HolderLookup;
-
-import java.util.HashSet;
-import java.util.Set;
-import java.util.concurrent.CompletableFuture;
-import java.util.stream.Stream;
 
 /**
  * 1.20.5+ variant of the language provider.
@@ -29,19 +29,19 @@ public class CreatureChatLangProvider extends FabricLanguageProvider {
 
     @Override
     public void generateTranslations(HolderLookup.Provider registryLookup, TranslationBuilder builder) {
-        Set<String> added = new HashSet<>();
+        Map<String, String> en = new TreeMap<>();
         Stream.of(
                 Randomizer.allErrorText(),
+                Randomizer.allNoResponseText(),
                 CCText.UI_TEXT.stream(),
                 CCText.CONFIG_TEXT.stream(),
                 EntityChatData.ERROR_MISC.stream(),
                 EntityChatData.ERROR_SOLUTIONS.stream(),
                 Advancements.allText()
-        ).flatMap(s -> s).forEach(tr -> {
-            if (added.add(tr.key())) {
-                builder.add(tr.key(), tr.en());
-            }
-        });
+        ).flatMap(s -> s).forEach(tr -> en.putIfAbsent(tr.key(), tr.en()));
+
+        en.forEach(builder::add);
+        LangSync.sync(en);
     }
 
     @Override